Format and fix PP W1

+++ Year 3/Semester 6/PP lab/Week 1/q2.c
@@ -1,21 +1,34 @@
+// Arithmetic operations
+
 #include <stdio.h>
 
-int main() {
+int main()
+{
 	int a, b;
 	char op;
-	
+
 	printf("Enter op, num, num\n");
 	scanf("%c", &op);
 	scanf("%d%d", &a, &b);
-	
-	switch (op) {
-		   case '+': printf("%d", a + b); break;
-		   case '-': printf("%d", a - b); break;
-		   case '*': printf("%d", a * b); break;
-		   case '/': printf("%d", a / b); break;
-		   default: printf("Wrong operator"); break;
+
+	switch (op)
+	{
+	case '+':
+		printf("%d", a + b);
+		break;
+	case '-':
+		printf("%d", a - b);
+		break;
+	case '*':
+		printf("%d", a * b);
+		break;
+	case '/':
+		printf("%f", (double)a / b);
+		break;
+	default:
+		printf("Wrong operator");
+		break;
 	}
-	
+
 	return 0;
 }
-
